+++ src/termview.c
@@ -895,9 +895,8 @@ _make_palette_from_color(s_termview *sd,
      }
    else
      {
-        const Eina_Bool true_colors = ! sd->nvim->opts->termcolors;
         const s_termview_color col =
-           termview_color_decompose((uint32_t)color, true_colors);
+           termview_color_decompose((uint32_t)color, sd->nvim->true_colors);
         return _make_palette(sd, col);
      }
 }